Some cleanup of structure
