+++ src/neo4j-arc/graph-visualization/GraphVisualizer/Graph/GraphEventHandlerModel.ts
@@ -17,25 +17,18 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { BasicNode, BasicNodesAndRels } from 'neo4j-arc/common'
 
-import { GraphModel } from './Graph'
-import { NodeModel } from './Node'
-import { RelationshipModel } from './Relationship'
-import { VizItem } from '../types'
+import { GraphModel } from '../../models/Graph'
+import { NodeModel } from '../../models/Node'
+import { RelationshipModel } from '../../models/Relationship'
+import { GetNodeNeighboursFn, VizItem } from '../../types'
 import {
   GraphStats,
   getGraphStats,
   mapNodes,
   mapRelationships
-} from '../utils/mapper'
-import { Visualization } from '../visualization/Visualization'
-
-export type GetNodeNeighboursFn = (
-  node: BasicNode | NodeModel,
-  currentNeighbourIds: string[],
-  callback: (data: BasicNodesAndRels) => void
-) => void
+} from '../../utils/mapper'
+import { Visualization } from './visualization/Visualization'
 
 export class GraphEventHandlerModel {
   getNodeNeighbours: GetNodeNeighboursFn
