+++ src/neo4j-arc/graph-visualization/GraphVisualizer/Graph/WheelZoomInfoOverlay.tsx
