+++ src/neo4j-arc/graph-visualization/GraphVisualizer/Graph/visualization/ForceSimulation.ts
@@ -39,11 +39,11 @@ import {
   PRECOMPUTED_TICKS,
   TICKS_PER_RENDER,
   VELOCITY_DECAY
-} from '../constants'
-import { GraphModel } from '../models/Graph'
-import { NodeModel } from '../models/Node'
-import { RelationshipModel } from '../models/Relationship'
-import circularLayout from '../utils/circularLayout'
+} from '../../../constants'
+import { GraphModel } from '../../../models/Graph'
+import { NodeModel } from '../../../models/Node'
+import { RelationshipModel } from '../../../models/Relationship'
+import circularLayout from './utils/circularLayout'
 
 const oneRelationshipPerPairOfNodes = (graph: GraphModel) =>
   Array.from(graph.groupedRelationships()).map(pair => pair.relationships[0])
