+++ src/neo4j-arc/graph-visualization/GraphVisualizer/Graph/visualization/renderers/d3Icons.ts
