+++ src/neo4j-arc/graph-visualization/GraphVisualizer/Graph/visualization/renderers/init.ts
@@ -19,8 +19,8 @@
  */
 import { BaseType } from 'd3-selection'
 
-import { NodeCaptionLine, NodeModel } from '../../models/Node'
-import { RelationshipModel } from '../../models/Relationship'
+import { NodeCaptionLine, NodeModel } from '../../../../models/Node'
+import { RelationshipModel } from '../../../../models/Relationship'
 import Renderer from '../Renderer'
 
 const noop = () => undefined
