+++ src/neo4j-arc/graph-visualization/GraphVisualizer/Graph/visualization/utils/circularLayout.ts
@@ -17,7 +17,7 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-import { NodeModel } from '../models/Node'
+import { NodeModel } from '../../../../models/Node'
 
 export default function circularLayout(
   nodes: NodeModel[],
