+++ src/neo4j-arc/graph-visualization/index.ts
@@ -17,11 +17,8 @@
  * You should have received a copy of the GNU General Public License
  * along with this program.  If not, see <http://www.gnu.org/licenses/>.
  */
-export { Graph } from './components/Graph'
-export type { GraphProps } from './components/Graph'
 
 export { GraphModel } from './models/Graph'
-export type { GetNodeNeighboursFn } from './models/GraphEventHandler'
 export { GraphStyleModel, Selector } from './models/GraphStyle'
 
 export type { NodeItem, RelationshipItem, VizItem } from './types'
